fix(docker): reduce docker image size
This reduces the size of the development and production images with approximately 50 percent.
Additionally, it enables a transition to a rootless container by ensuring permissions of data folders
are set to allow the www-data user to perform operations.

This is a BREAKING change: the location of the data mount has changed.
Not changing the mount will result in the logs no longer being stored.
With these changes implemented, replenish has reduced complexity (it only needs to worry about
the public folder). Cache etc. is stored outside of the volumes.